requirements.txt
Pillow
reportlab
PyMuPDF
streamlit
